libalpm: Properly cast int to size_t
The vsnprintf call returns an int and the memory allocation tries to
allocate as many bytes as vsnprintf would have printed + 1. If vsnprintf
returns exactly INT_MAX, a signed integer overflow would occur, which is
undefined behavior.

Most likely, the eventual SIZE_MAX byte allocation simply fails, and it
is highly unlikely that the privileged process would try to write such a
log.

But let's better be safe than sorry.

+++ lib/libalpm/sandbox.c
@@ -141,8 +141,8 @@ void _alpm_sandbox_cb_log(void *ctx, alpm_loglevel_t level, const char *fmt, va_
 	if(string_size <= 0) {
 		return;
 	}
-	MALLOC(string, string_size + 1, return);
-	string_size = vsnprintf(string, string_size + 1, fmt, args);
+	MALLOC(string, (size_t)string_size + 1, return);
+	string_size = vsnprintf(string, (size_t)string_size + 1, fmt, args);
 	if(string_size > 0) {
 		write_to_pipe(context->callback_pipe, &type, sizeof(type));
 		write_to_pipe(context->callback_pipe, &level, sizeof(level));
